fix: rules are at top level (not under overrides)
BREAKING CHANGE: the rules are provided at the top level,
instead of under an `overrides` property.
Providing the rules under the `overrides` property was never a good idea.
It prevents specifying which files the rules apply to (e.g. `[*.js, *.ts]`).
I apologize.
To migrate, make sure that your `extends` property is under an [`overrides` item][overrides].
An example is in the readme.
To help verify your configuration,
you could obtain a list of files that will be linted, this way:
`DEBUG=eslint:cli-engine npx eslint <path>`.

[overrides]: https://eslint.org/docs/latest/use/configure/configuration-files#how-do-overrides-work

closes #1149
closes #1088

Co-authored-by: Rostislav Simonik <rostislav.simonik@technologystudio.sk>